\name{aaresponse-package}
\alias{aaresponse-package}
\alias{aaresponse}
\docType{package}
\title{
\packageTitle{aaresponse}
}
\description{
\packageDescription{aaresponse}
}
\author{
\packageAuthor{aaresponse}

Maintainer: \packageMaintainer{aaresponse}
}
\reference{R. Wehrens, J. Engel, J. Mes, A. de Jong and
  D. Esser. Analysing postprandial amino acid responses in crossover
  studies with the aaresponse package for R. Amino Acids (2024) 65:25.}
\keyword{ package }
