\name{supro}
\alias{supro}
\docType{data}
\title{Data from an experiment comparing the nutritional value of three
different proteins: WP (whey protein), BP (bovine plasma protein) and CP
(corn protein).}
\description{
A data set from an experiment in which 12 participants were administered
three different protein meals (after fasting) in three consecutive weeks
according to a crossover design. Levels of 19 amino acids in the blood
of these participants were determined at regular time points after each
meal. Two totals are included as well: total amino acids (TAA), and
total of essential amino acids (TEAA).}
\source{
D. Esser et al. Manuscript in preparation, 2022.
}
\examples{
data(supro)
head(supro)
}
\keyword{datasets}
